pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "self-media-workbench"
version = "0.1.0"
description = "Local-first content discovery, fetching, and writing workbench."
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.12",
  "PyYAML>=6.0",
  "requests>=2.31",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[tool.pytest.ini_options]
testpaths = [
  "tests",
]
pythonpath = [
  "scripts",
]
addopts = "-q --import-mode=importlib --basetemp=runtime/pytest-tmp"
